requirements.txt
# OmniDupe - Duplicate Image Finder
# Python Dependencies

# Core image processing
Pillow>=10.0.0

# Image metadata extraction
ExifRead>=3.0.0

# Perceptual hashing
ImageHash>=4.3.1

# Database (included in Python standard library)
# sqlite3

# Standard library dependencies used:
# - argparse (command line parsing)
# - logging (application logging)
# - pathlib (path handling)
# - sqlite3 (database operations)
# - concurrent.futures (parallel processing)
# - hashlib (file hashing)
# - mimetypes (MIME type detection)
# - os (operating system interface)
# - tempfile (temporary files)
# - datetime (date/time handling)

# Development and testing dependencies (optional)
# pytest>=8.0.0
# pytest-cov>=6.0.0
